GUI started for user options
Added JMenuBar which presents two options (Options, and Restart
Simulation). The Options option will let the user set the simulation
constants. At the moment it just sets the ocean width and depth so we
can preview it working. The simulation must be restarted after any
changes to the options. This is was the next option is for.