Breeding Bug Fix
Fixed bug where creature list cannot be added to due to
ConcurrentModificationException. Changed breed so that it now is public
and is called within the Simulator simulateOneStep iteration. Also
changed the simulateOneStep method to use an Iterator since that way we
can add to the list while it is being iterated and avoid the exception
above.