Major Bug Fix
The simulation appeared to not be giving consistent results across IDE's
and other JVMs. This is because a HashMap was being used within the
creationDecider. This was a problem because a HashMap does NOT guarantee
order of elements. Therefore it was necessary to use a LinkedHashMap.
Also moved the LinkedHashMap to be an instance field like it should be
and being instantiated in the constructor.